Added class DeactivableItem to group all items that can be deactivated
