+++ app/src/main/java/ch/epfl/sweng/SDP/game/drawing/items/BumpingItem.java
@@ -62,11 +62,6 @@ public class BumpingItem extends Item {
     }
 
     @Override
-    public void deactivate(PaintView paintView) {
-        // Must be implemented, but does nothing.
-    }
-
-    @Override
     public String getTextFeedback() {
         return " ";
     }
